Sum Channel Block DSP Logic Design
The block is used to maintain the history of symbols in a DSP system. It provides the sum of the last 4 or 8 valid values, coming to the block from 64 input channels. The sum goes to the output 5 clocks after data has arrived, including the latest valid data on data_in. The system also provides valid sum even if only one valid data item is received for a channel.

The input data is 8 bit wide and is treated as unsigned.